Split magnetometer node into C++ class and header
